+++ client/src/components/wizard/Wizard.tsx
@@ -6,11 +6,19 @@ interface WizardProps {
   handleSelection: (text: string) => void;
   open: boolean;
   onClose: () => void;
+  showFakeData?: boolean;
+  showTemplates?: boolean;
+  showPreviousSteps?: boolean;
+  showAi?: boolean;
 }
 export const Wizard: React.FC<WizardProps> = ({
   handleSelection,
   open,
   onClose,
+  showFakeData,
+  showAi,
+  showPreviousSteps,
+  showTemplates,
 }) => {
   return (
     <Modal
@@ -19,7 +27,14 @@ export const Wizard: React.FC<WizardProps> = ({
       onClose={onClose}
     >
       <Paper className="flex flex-col w-96 rounded-sm outline-none p-4">
-        <WizardMenu handleSelection={handleSelection} onClose={onClose} />
+        <WizardMenu
+          handleSelection={handleSelection}
+          onClose={onClose}
+          showAi={showAi}
+          showFakeData={showFakeData}
+          showPreviousSteps={showPreviousSteps}
+          showTemplates={showTemplates}
+        />
       </Paper>
     </Modal>
   );
