+++ client/src/components/wizard/WizardMenu.tsx
@@ -1,15 +1,27 @@
 import React, { useState } from "react";
 import { WizardItem } from "./WizardItem";
 import { WizardTemplate } from "./WizardTemplate";
-import { FakeDataWizard, TemplateWizard } from "./FakeDataWizard";
+import {
+  FakeDataWizard,
+  PreviousStepsWizard,
+  TemplateWizard,
+} from "./FakeDataWizard";
 
 interface WizardMenuProps {
   handleSelection: (text: string) => void;
   onClose: () => void;
+  showFakeData?: boolean;
+  showTemplates?: boolean;
+  showPreviousSteps?: boolean;
+  showAi?: boolean;
 }
 export const WizardMenu: React.FC<WizardMenuProps> = ({
   handleSelection,
   onClose,
+  showFakeData = true,
+  showTemplates = true,
+  showPreviousSteps = true,
+  showAi = true,
 }) => {
   const [wizardType, setWizardType] = useState("");
 
@@ -17,15 +29,25 @@ export const WizardMenu: React.FC<WizardMenuProps> = ({
     <>
       {wizardType === "" && (
         <WizardTemplate onClose={onClose} title="Generate Data">
-          <WizardItem
-            title="Fake Data"
-            onClick={() => setWizardType("fake-data")}
-          />
-          <WizardItem
-            title="Templates"
-            onClick={() => setWizardType("templates")}
-          />
-          <WizardItem title="AI ⭒˚" />
+          {showFakeData && (
+            <WizardItem
+              title="Fake Data"
+              onClick={() => setWizardType("fake-data")}
+            />
+          )}
+          {showTemplates && (
+            <WizardItem
+              title="Templates"
+              onClick={() => setWizardType("templates")}
+            />
+          )}
+          {showPreviousSteps && (
+            <WizardItem
+              title="Previous Steps"
+              onClick={() => setWizardType("previous-steps")}
+            />
+          )}
+          {showAi && <WizardItem title="AI ⭒˚" />}
         </WizardTemplate>
       )}
       {wizardType === "fake-data" && (
@@ -37,6 +59,13 @@ export const WizardMenu: React.FC<WizardMenuProps> = ({
       )}
       {wizardType === "templates" && (
         <TemplateWizard
+          onClose={onClose}
+          goBack={() => setWizardType("")}
+          handleSelection={handleSelection}
+        />
+      )}
+      {wizardType === "previous-steps" && (
+        <PreviousStepsWizard
           onClose={onClose}
           goBack={() => setWizardType("")}
           handleSelection={handleSelection}
